Complete integration of NeurologyTable into patient form
I switched from using global variables for the React components to using
RequireJS. I needed to do this because XW uses RequireJS, and React
doesn't attach globally if RequireJS is present. This is a much cleaner
solution anyway.
I also moved from using a Browserify-generated bundle.js to using
WebJars with RequireJS. This is better suited to XW and is also better
with respect to possible vendor JS conflicts.