Complete persistence of NeurologyTable data
Persistence is achieved using a JSON object passed between frontend and
backend. The backend listens for document updates where the JSON object
is present, passed in a hidden form field.
I wrote a controller for this task, expecting that I would be able to
use it in the event listener, but did not use it because it caused
infinite recursion in the event handler. I've kept it here so that it
can be used for JSON export etc.